Fixed elf64 objects on TB and minor linux compat issues on windows, increased column limit to 32bits

+++ include/cuik.h
@@ -148,8 +148,8 @@ typedef struct SourceLine {
 
 typedef struct SourceLoc {
     SourceLine* line;
-    short columns;
-    short length;
+    unsigned int columns;
+    unsigned int length;
 } SourceLoc;
 
 typedef struct Cuik_FileEntry {
